Refac: 불필요 Layer(DefaultCarFactory) 제거
