+++ src/main/kotlin/carRacing/serviceimpl/RandomMovementCarFactory.kt
@@ -1,15 +1,19 @@
 package carRacing.serviceimpl
 
 import carRacing.domain.Car
+import carRacing.service.CarFactory
 import carRacing.view.CarInfo
 
 /*
 * RandomMovementController를 주입하는 CarFactory
 * */
-class RandomMovementCarFactory : DefaultCarFactory() {
+class RandomMovementCarFactory : CarFactory {
+
+    override fun createCars(carInfoList: List<CarInfo>): List<Car> = carInfoList.map { info -> Car(info.name, info.movementController) }
+
     fun createCarList(carNameList: List<String>): List<Car> {
         val carInfoList: List<CarInfo> = carNameList.map { CarInfo(it, RandomMovementController()) }
 
-        return super.createCars(carInfoList)
+        return createCars(carInfoList)
     }
 }
